core/requirements.txt
# Data
yfinance==0.2.40
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
beautifulsoup4==4.12.3
feedparser==6.0.11

# AI
anthropic==0.29.0

# Finance (KRX)
pykrx==1.0.45

# Utils
python-dotenv==1.0.1
loguru==0.7.2
schedule==1.2.2
